refactor: reorganize auth fixtures and update readme
