+++ .env_example.txt
@@ -17,3 +17,5 @@ ALCHEMY_ECHO='True'
 
 USERNAME='admin'
 PASSWORD='moon-rover-secret'
+
+OBSTACLES_JSON_PATH='./config/obstacles.json'
