+++ .gitignore
@@ -212,4 +212,5 @@ __marimo__/
 .docs 
 
 Features.md
+TODO.md
