+++ app/infrastructure/repositories/auth_provider.py
@@ -4,7 +4,7 @@ from pydantic_settings import BaseSettings, SettingsConfigDict
 class BasicAuthSettings(BaseSettings):
     """Basic auth configuration from environment"""
 
-    USERNAME: str = 'admin'
-    PASSWORD: str = 'moon-rover-secret'
+    USERNAME: str
+    PASSWORD: str 
 
     model_config = SettingsConfigDict(extra='ignore')
